+++ solvers/__init__.py
@@ -8,5 +8,6 @@ from .gurobi import GurobiSolver
 from .ortools import ORToolsSolver
 from .cbc import CBCSolver
 from .scip import SCIPSolver
+from .deap import DEAPSolver
 from .localsearch import LocalSearchSolver
 from .tabusearch import TabuSearchSolver