+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feature - Added 3 new alternative solvers for the system.
 - Feature - Added SCIPSolver for alternative solution.
 - Feature - Added Local Search and Tabu Search solvers for alternative solution.
+- Feature - Added DEAPSolver for alternative solution.
 
 ### Changed
 - Updated statistics visualization buttons and their information display options.